+++ src/components/locale/cultures/ar-SA.js
@@ -22,8 +22,10 @@ Soho.Locale.addCulture('ar-SA', {
         year: 'MMMM yyyy',
         dayOfWeek: 'EEE d',
         timestamp: 'h:mm:ss a',
+        timestampMillis: 'h:mm:ss.SSS a',
         hour: 'h:mm a',
         datetime: 'yyyy/MM/dd h:mm a',
+        datetimeMillis: 'yyyy/MM/dd h:mm:ss.SSS a',
         timezone: 'yyyy/MM/dd h:mm a zz',
         timezoneLong: 'yyyy/MM/dd h:mm a zzzz'
       }, // Infered short + short gregorian/dateTimeFormats